pyproject.toml
[tool.poetry]
name = "nonebot-plugin-hammer-core"
version = "0.3.0"
description = "core dependency of nonebot-plugin-hammer-xxx"
authors = ["ArgonarioD <739062975@qq.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://docs.hammer-hfut.tk:233"
repository = "https://github.com/ArgonarioD/nonebot-plugin-hammer-core"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9"
]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.9"
nonebot-adapter-onebot = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

